# Faster compilation with maps instead of function heads