--- a/lib/elixir/unicode/unicode.ex
+++ b/lib/elixir/unicode/unicode.ex
@@ -309,18 +309,17 @@ defmodule String.Graphemes do
 
   # Handle Hangul L
 
-  @cluster_reverse_map (for {class, list} <- cluster, codepoint <- list do
-    {codepoint, class}
-  end
-  |> Map.new)
+  @cluster_reverse_map (for {class, list} <- cluster, class == "L" or class == "LV" or class == "LVT", codepoint <- list, into: %{} do
+    {codepoint, String.to_atom(class)}
+  end)
 
-  defp next_hangul_l_size(string, size) do
-    {codepoint, rest} = String.next_codepoint(string)
-    class_for_codepoint = @cluster_reverse_map[codepoint]
+  defp next_hangul_l_size(<<unicode_cp::utf8, rest::binary>> = string, size) do
+    codepoint = <<unicode_cp::utf8>>
+    class_for_codepoint = Map.get(@cluster_reverse_map, codepoint)
     case class_for_codepoint do
-      "L" ->   next_hangul_l_size(rest, size + byte_size(codepoint))
-      "LV" ->  next_hangul_v_size(rest, size + byte_size(codepoint))
-      "LVT" -> next_hangul_t_size(rest, size + byte_size(codepoint))
+      :L ->   next_hangul_l_size(rest, size + byte_size(codepoint))
+      :LV ->  next_hangul_v_size(rest, size + byte_size(codepoint))
+      :LVT -> next_hangul_t_size(rest, size + byte_size(codepoint))
       _ ->     next_hangul_v_size(string, size)
     end
   end