Speed up compilation of the unicode module by switching `next_hangul_l_size`
from using many function heads to doing lookups in a map.

As of Elixir 1.2.3 / Erlang 18, map performance is much improved. In my
experiments, in cases where we associate static keys and values, looking up a
key in a map performs just as well as finding a matching function head for the
static key, but compilation time grows much more slowly as the number of keys
increases.

Compile time is obviously much less important than run time performance, however. Here's what and how I've measured - **please tell me if I'm missing something**:

- `time (cd lib/elixir &&
 \ ../../bin/elixirc --verbose --ignore-module-conflict unicode/unicode.ex -o ebin)`
 \ to measure compile time on both master and this feature branch
- `elixir speed_test.exs` to see the runtime for calculating string graphemes

My results were:

     -------------------------------------------------------
     |                        | master      | PR           |
     -------------------------------------------------------
     | compile `unicode.ex`   | 23s         | 9s           |
     -------------------------------------------------------
     | `speed_test.exs`       | 0.6 μs/call | 0.6 μs/call  |
     -------------------------------------------------------
     | mem for speed test     | 30MB        | 30 MB        |
     -------------------------------------------------------

I used this script for timing the runtime execution:

     # speed_test.exs
     IO.inspect ["working correctly?", String.graphemes("\u1100\u115D\uB4A4") == ["ᄀᅝ뒤"]]

     some_func = fn ->
       String.graphemes("\u1100\u115D\uB4A4")
     end

     run_n_times = fn (count, func) ->
       Enum.each(1..count, fn (_i) ->
         func.()
       end)
     end

     n = 800_000

     {microseconds, :ok} = :timer.tc(run_n_times, [n, some_func])
     IO.puts "#{microseconds / n} microseconds per call (#{microseconds} total for #{n} calls)"

See also: [an earlier (and simpler) experiment I tried comparing maps and function heads](https://gist.github.com/nathanl/27e19896bba489799804).