@nathanl 2016-03-16
I originally planned to try switching the Unicode case conversion, but it wasn't actually the slow part of compiling this file.