@ericmj 2016-03-16
I bet this version will produce a smaller file size also. Can you compare the size of `Elixir.String.Unicode.beam`?