@ericmj 2016-03-16
> I'm afraid proposed implementation generates more garbage binaries since we're using String.next_codepoint but previously it should have been optimized.

We can work around this by pattern matching in the function head.