@lexmag 2016-03-16
> We can work around this by pattern matching in the function head.

As I see we can't turn `next_grapheme` call into pattern.