@nathanl 2016-03-16
I think `next_codepoint` can't be done as a pattern match because the length varies.

I don't understand what's meant by "garbage binaries" - can somebody explain?