@nathanl 2016-03-16
@ericmj Re file sizes:

branch
216K Elixir.String.Graphemes.beam
169K Elixir.String.Normalizer.beam
192K Elixir.String.Unicode.beam

master
153K Elixir.String.Graphemes.beam
169K Elixir.String.Normalizer.beam
192K Elixir.String.Unicode.beam