@ericmj 2016-03-16
> As I see we can't turn next_grapheme call into pattern.

That function is not being called.

> I think next_codepoint can't be done as a pattern match because the length varies.

You can - check the implementation of that function in the same module.