@lexmag 2016-03-16
> I don't understand what's meant by "garbage binaries" - can somebody explain?

Currently we have optimization: all the recursive `next_hangul_l_size` calls use references to the same binary. With this PR we start to allocate new binary every `next_hangul_l_size` call.