@nathanl 2016-03-16
Came back from dinner to say - I think I measured the speed incorrectly. Please ignore until I can confirm/deny. :frowning: