@josevalim 2016-03-16
Thanks @nathanl! Regarding the benchmarks, also please make sure you start each measurement in a new process to avoid GC interference. :)