@nathanl 2016-03-18
OK, I was a total doofus and wasn't running my actual compiled version of Elixir when testing. :laughing:

Corrected results with latest changes:

```
-------------------------------------------------------
|                        | master      | PR           |
-------------------------------------------------------
| compile `unicode.ex`   | 27s         | 10s          |
-------------------------------------------------------
| `speed_test.exs`       | 0.6 μs/call | 1.2 μs/call  |
-------------------------------------------------------
| mem for speed test     | 29.7M       | 31.1 MB      |
-------------------------------------------------------
```

So on my machine, this saves 17 seconds of compile time but adds 0.2 microseconds per hangul_l grapheme and ~1.5MB of memory usage when `String.graphemes` is used.