@nathanl 2016-03-18
>  Regarding the benchmarks, also please make sure you start each measurement in a new process to avoid GC interference.

I ran `bin/elixir speed_test.exs` each time (see code above).

> Compile time is obviously much less important than run time performance, however.

I said this, but in this case, the run time hit is only for hangul characters. What concerns me more is the difference in memory usage that comes from saving the map as an attribute.

What do you all think? Is this a good or bad tradeoff?