@josevalim 2016-03-18
Thanks @nathanl. I think the trade-offs with this particular solution is not worth it but I think this is definitely an area worth exploring more. Maybe we can find better ways of compiling the code? For example, something that relies more on guards and ranges than on binary match. For example, if the hangul codepoints are all in a row, we could compile it to:

    def hangul_l(<<h::utf8, t::binary>>, _) when h in 0x3210..0x3ad3 do

Or even:

    def hangul_l(<<h::utf8, t::binary>>, _) when h in [0x3210, ..., 0x3ad3] do

This should reduce the size and not have such a large impact on performance. Could you try exploring those alternative solutions?