@nathanl 2016-03-18
> Could you try exploring those alternative solutions?

Yep, happy to. Thanks for the ideas and for supporting my tinkering! :smile: 

> This should reduce the size and not have such a large impact on performance.

Can you give me any sense of what is or isn't an acceptable performance impact for you?