@josevalim 2016-03-18
@nathanl I am not sure. Twice slower is quite the cost, specially with operations that are linear. Btw, I did a quick check and it seems LVT, which are the majority of hangul codepoints, is a single very big range. It is very likely we can compile it to a single function clause, which would optimize this quite considerably and likely be faster at runtime too.