@josevalim 2016-03-18
I have done a very quick check to explore these solutions. Today we generate one clause per codepoint. If we compile them to ranges, here is the result we get:

```
CR+LF+Control
{23, 3984}
L
{2, 125}
T
{2, 137}
LV
{399, 399}
LVT
{399, 10773}
Extend+SpacingMark
{415, 1940}
Regional_Indicator
{1, 26}
```

For each group, the number of new clauses (range-based) is shown on the left, the number of old clauses on the right (codepoint-based).