@josevalim 2016-03-18
@nathanl Based on this range idea, I have pushed an initial work here: https://github.com/elixir-lang/elixir/tree/jv-unicode-range

I have currently replaced all clauses in `next_grapheme_size` by range clauses, however, I have not benchmarked nor evaluated the results in anyway. Could you please benchmark and let us know of your findings?

For example, if it makes everything slower, we can consider using this approach only on part of the next_grapheme_size functions, like the hangul ones and that's it. Looking forward to your results, thank you!