@nathanl 2016-03-18
@josevalim This is what I got (with [a tiny change](https://github.com/elixir-lang/elixir/pull/4404)):

```
-------------------------------------------------------
|                        | e8da279     | eff6d30      |
-------------------------------------------------------
| compile `unicode.ex`   | 30s         | 12s          |
-------------------------------------------------------
| `speed_test.exs`       | 0.6 μs/call | 1.1 μs/call  |
-------------------------------------------------------
| mem for speed test     | 29.6M       | 29.6 MB      |
-------------------------------------------------------

```

What I do is run `./go.sh`....


```
#!/bin/bash
# go.sh
echo "compiling" &&
if [ -f lib/elixir/unicode/utils.ex ]; then
  time (cd lib/elixir && ../../bin/elixirc --verbose --ignore-module-conflict -r unicode/utils.ex unicode/unicode.ex -o ebin)
else
  time (cd lib/elixir && ../../bin/elixirc --verbose --ignore-module-conflict unicode/unicode.ex -o ebin)
fi
```

Then run this `bin/elixir speed_test.exs` several times, with `watch -n 0.1 ps -f -o rss ` running in another terminal to see the memory consumption (probably not super accurate):


```
# speed_test.exs
some_func = fn -> 
  String.graphemes("\u1100\u115D\uB4A4")
end

run_n_times = fn (count, func) ->
  Enum.each(1..count, fn (_i) ->
    func.()
  end)
end

# n = 20
n = 800_000

{microseconds, :ok} = :timer.tc(run_n_times, [n, some_func])
IO.puts "#{microseconds / n} microseconds per call (#{microseconds} total for #{n} calls)"

:timer.sleep(3_000)
```