@nathanl 2016-03-18
I don't really understand why your version is slower. The VM has fewer function heads to look through, and I'd expect a range inclusion check to be fast.