@josevalim 2016-03-18
That's awesome, thank you! I will play with those benchmarks as well. So it seems to be slightly faster than your previous version but overall slower. The good news is that Erlang 19 is going to make those cases faster, so I will benchmark on there too.